+++ android/app/src/main/kotlin/id/go/jabarprov/pikobar/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.pikobar_flutter
+package id.go.jabarprov.pikobar
 
 import android.os.Bundle
 import io.flutter.app.FlutterActivity
